MapSelect -> MapSelectGeneric
MapSelectAmsterdam -> MapSelect
