+++ src/components/MapSelectGeneric/style.scss
