+++ src/signals/incident/components/IncidentPreview/components/MapSelect/index.js
@@ -3,31 +3,33 @@ import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { themeSpacing } from '@amsterdam/asc-ui';
 
-import MapSelect from 'components/MapSelect';
-import configuration from 'shared/services/configuration/configuration';
 import { incidentType } from 'shared/types';
 
-export const getLatlng = location =>
-  location?.geometrie?.coordinates
-    ? {
-      latitude: location.geometrie.coordinates[1],
-      longitude: location.geometrie.coordinates[0],
-    }
-    : {
-      latitude: configuration.map.options.center[0],
-      longitude: configuration.map.options.center[1],
-    };
+import MapSelect from 'components/MapSelect';
+import { getOVLIcon } from '../../../form/MapSelect/iconMapping';
+
+export const DEFAULT_COORDS = [4.900312721729279, 52.37248465266875];
+
+export const getLatlng = location => {
+  const coords = location?.geometrie?.coordinates || DEFAULT_COORDS;
+  return {
+    latitude: coords[1],
+    longitude: coords[0],
+  };
+};
 
 const Values = styled.div`
   margin-bottom: ${themeSpacing(4)};
 `;
 
-const MapSelectPreview = ({ value, meta, incident }) => (
+const MapSelectPreview = ({ value, endpoint, incident }) => (
   <Fragment>
     <Values>{value.join('; ')}</Values>
     <MapSelect
-      geojsonUrl={meta.endpoint}
-      idField={meta.idField}
+      geojsonUrl={endpoint}
+      getIcon={getOVLIcon}
+      iconField="type_name"
+      idField="objectnummer"
       latlng={getLatlng(incident.location)}
       selectionOnly
       value={value}
@@ -37,10 +39,7 @@ const MapSelectPreview = ({ value, meta, incident }) => (
 
 MapSelectPreview.propTypes = {
   incident: incidentType,
-  meta: PropTypes.shape({
-    endpoint: PropTypes.string.isRequired,
-    idField: PropTypes.string.isRequired,
-  }),
+  endpoint: PropTypes.string.isRequired,
   value: PropTypes.arrayOf(PropTypes.string),
 };
