+++ src/signals/incident/components/IncidentPreview/components/index.js
@@ -2,7 +2,7 @@ import DateTime from './DateTime';
 import Image from './Image';
 import Map from './Map';
 import MapSelectPreview from './MapSelect';
-import MapSelectAmsterdamPreview from './MapSelectAmsterdam';
+import MapSelectGenericPreview from './MapSelectGeneric';
 import ListObjectValue from './ListObjectValue';
 
 export default {
@@ -11,5 +11,5 @@ export default {
   Image,
   Map,
   MapSelectPreview,
-  MapSelectAmsterdamPreview,
+  MapSelectGenericPreview,
 };
