+++ src/signals/incident/components/IncidentPreview/components/index.test.js
@@ -2,7 +2,7 @@ import DateTime from './DateTime';
 import Image from './Image';
 import Map from './Map';
 import MapSelectPreview from './MapSelect';
-import MapSelectAmsterdamPreview from './MapSelectAmsterdam';
+import MapSelectGenericPreview from './MapSelectGeneric';
 
 import components from '.';
 import ListObjectValue from './ListObjectValue';
@@ -15,7 +15,7 @@ describe('Preview components', () => {
       Image,
       Map,
       MapSelectPreview,
-      MapSelectAmsterdamPreview,
+      MapSelectGenericPreview,
     });
   });
 });
