+++ src/signals/incident/components/form/MapSelect/index.js
@@ -1,30 +1,32 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import get from 'lodash.get';
 import styled from 'styled-components';
 import { themeSpacing } from '@amsterdam/asc-ui';
 
 import MapSelectComponent from 'components/MapSelect';
-import configuration from 'shared/services/configuration/configuration';
 
 import Header from '../Header';
+import { getOVLIcon, LEGEND_ITEMS } from './iconMapping';
 
-export const getLatlng = meta =>
-  meta?.incidentContainer?.incident?.location?.geometrie?.coordinates
-    ? {
-      latitude: meta.incidentContainer.incident.location.geometrie.coordinates[1],
-      longitude: meta.incidentContainer.incident.location.geometrie.coordinates[0],
-    }
-    : {
-      latitude: configuration.map.options.center[0],
-      longitude: configuration.map.options.center[1],
-    };
+const filter_legend = (items, types) => items.filter(element => types.includes(element.key));
+
+const DEFAULT_COORDS = [4.900312721729279, 52.37248465266875];
+
+const getLatlng = meta => {
+  const coords = get(meta, 'incidentContainer.incident.location.geometrie.coordinates', DEFAULT_COORDS);
+  return {
+    latitude: coords[1],
+    longitude: coords[0],
+  };
+};
 
 const Selection = styled.span`
   display: inline-block;
   margin-top: ${themeSpacing(3)};
 `;
 
-const MapSelect = ({ handler, touched, hasError = () => {}, meta, parent, getError, validatorsOrOpts }) => {
+const MapSelect = ({ handler, touched, hasError, meta, parent, getError, validatorsOrOpts }) => {
   const onSelectionChange = selection => {
     const value = [...selection.set.values()];
     parent.meta.updateIncident({ [meta.name]: value });
@@ -32,6 +34,7 @@ const MapSelect = ({ handler, touched, hasError = () => {}, meta, parent, getErr
 
   const latlng = getLatlng(parent.meta);
   const url = meta.endpoint;
+  const filtered_legend = filter_legend(LEGEND_ITEMS, meta.legend_items);
 
   // Get selection array from "handler".
   // the value is not always an array (it's a string on load).
@@ -52,21 +55,26 @@ const MapSelect = ({ handler, touched, hasError = () => {}, meta, parent, getErr
       >
         <MapSelectComponent
           geojsonUrl={url}
+          getIcon={getOVLIcon}
           hasGPSControl
-          idField={meta.idField}
+          iconField="type_name"
+          idField="objectnummer"
           latlng={latlng}
+          legend={filtered_legend}
           onSelectionChange={onSelectionChange}
           value={selection}
           zoomMin={meta.zoomMin}
         />
         {selection.length > 0 && (
-          <Selection>
-            Het gaat om{meta.selectionLabel ? ` ${meta.selectionLabel}` : ''}: {selection.join('; ')}
-          </Selection>
+          <Selection>Het gaat om lamp of lantaarnpaal met nummer: {selection.join('; ')}</Selection>
         )}
       </Header>
     )
   );
+};
+
+MapSelect.defaultProps = {
+  hasError: () => {},
 };
 
 MapSelect.propTypes = {
@@ -76,10 +84,9 @@ MapSelect.propTypes = {
   meta: PropTypes.shape({
     className: PropTypes.string,
     endpoint: PropTypes.string.isRequired,
-    idField: PropTypes.string.isRequired,
     isVisible: PropTypes.bool,
+    legend_items: PropTypes.arrayOf(PropTypes.string.isRequired).isRequired,
     name: PropTypes.string,
-    selectionLabel: PropTypes.string,
     zoomMin: PropTypes.number,
   }),
   parent: PropTypes.object,
