--- src/signals/incident/components/form/MapSelectAmsterdam/iconMapping.test.js
@@ -1,26 +1,0 @@
-import { OVL_MAPPING, getOVLIcon, LEGEND_ITEMS } from './iconMapping';
-
-describe('getOVLIcon', () => {
-  it('should get default icon', () => {
-    expect(getOVLIcon('Klok', false)).toBe(OVL_MAPPING.Klok.default);
-  });
-
-  it('should get select icon', () => {
-    expect(getOVLIcon('Klok', true)).toBe(OVL_MAPPING.Klok.selected);
-  });
-
-  it('should default to first icon if missing', () => {
-    jest.spyOn(global.console, 'error').mockImplementation(() => {});
-    expect(getOVLIcon('missing', false)).toBe(OVL_MAPPING.Klok.default);
-    expect(global.console.error).toHaveBeenCalledWith('icon missing for type, using default. Type is: missing');
-  });
-
-  it('should have matching keys', () => {
-    const mappingKeys = Object.keys(OVL_MAPPING).map(key => key.toLowerCase());
-    const legendKeys = LEGEND_ITEMS.map(({ key }) => key);
-
-    mappingKeys.forEach(mappingKey => {
-      expect(legendKeys.includes(mappingKey));
-    });
-  });
-});
