+++ src/signals/incident/components/form/MapSelectGeneric/index.js
@@ -1,32 +1,30 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import get from 'lodash.get';
 import styled from 'styled-components';
 import { themeSpacing } from '@amsterdam/asc-ui';
 
-import MapSelectComponent from 'components/MapSelectAmsterdam';
+import MapSelectComponent from 'components/MapSelectGeneric';
+import configuration from 'shared/services/configuration/configuration';
 
 import Header from '../Header';
-import { getOVLIcon, LEGEND_ITEMS } from './iconMapping';
 
-const filter_legend = (items, types) => items.filter(element => types.includes(element.key));
-
-const DEFAULT_COORDS = [4.900312721729279, 52.37248465266875];
-
-const getLatlng = meta => {
-  const coords = get(meta, 'incidentContainer.incident.location.geometrie.coordinates', DEFAULT_COORDS);
-  return {
-    latitude: coords[1],
-    longitude: coords[0],
-  };
-};
+export const getLatlng = meta =>
+  meta?.incidentContainer?.incident?.location?.geometrie?.coordinates
+    ? {
+      latitude: meta.incidentContainer.incident.location.geometrie.coordinates[1],
+      longitude: meta.incidentContainer.incident.location.geometrie.coordinates[0],
+    }
+    : {
+      latitude: configuration.map.options.center[0],
+      longitude: configuration.map.options.center[1],
+    };
 
 const Selection = styled.span`
   display: inline-block;
   margin-top: ${themeSpacing(3)};
 `;
 
-const MapSelectAmsterdam = ({ handler, touched, hasError, meta, parent, getError, validatorsOrOpts }) => {
+const MapSelectGeneric = ({ handler, touched, hasError = () => {}, meta, parent, getError, validatorsOrOpts }) => {
   const onSelectionChange = selection => {
     const value = [...selection.set.values()];
     parent.meta.updateIncident({ [meta.name]: value });
@@ -34,7 +32,6 @@ const MapSelectAmsterdam = ({ handler, touched, hasError, meta, parent, getError
 
   const latlng = getLatlng(parent.meta);
   const url = meta.endpoint;
-  const filtered_legend = filter_legend(LEGEND_ITEMS, meta.legend_items);
 
   // Get selection array from "handler".
   // the value is not always an array (it's a string on load).
@@ -46,7 +43,7 @@ const MapSelectAmsterdam = ({ handler, touched, hasError, meta, parent, getError
     meta?.isVisible && (
       <Header
         // className value is referenced by form component
-        className="mapSelect"
+        className="mapSelectGeneric"
         meta={meta}
         options={validatorsOrOpts}
         touched={touched}
@@ -55,38 +52,34 @@ const MapSelectAmsterdam = ({ handler, touched, hasError, meta, parent, getError
       >
         <MapSelectComponent
           geojsonUrl={url}
-          getIcon={getOVLIcon}
           hasGPSControl
-          iconField="type_name"
-          idField="objectnummer"
+          idField={meta.idField}
           latlng={latlng}
-          legend={filtered_legend}
           onSelectionChange={onSelectionChange}
           value={selection}
           zoomMin={meta.zoomMin}
         />
         {selection.length > 0 && (
-          <Selection>Het gaat om lamp of lantaarnpaal met nummer: {selection.join('; ')}</Selection>
+          <Selection>
+            Het gaat om{meta.selectionLabel ? ` ${meta.selectionLabel}` : ''}: {selection.join('; ')}
+          </Selection>
         )}
       </Header>
     )
   );
 };
 
-MapSelectAmsterdam.defaultProps = {
-  hasError: () => {},
-};
-
-MapSelectAmsterdam.propTypes = {
+MapSelectGeneric.propTypes = {
   handler: PropTypes.func.isRequired,
   touched: PropTypes.bool,
   hasError: PropTypes.func,
   meta: PropTypes.shape({
     className: PropTypes.string,
     endpoint: PropTypes.string.isRequired,
+    idField: PropTypes.string.isRequired,
     isVisible: PropTypes.bool,
-    legend_items: PropTypes.arrayOf(PropTypes.string.isRequired).isRequired,
     name: PropTypes.string,
+    selectionLabel: PropTypes.string,
     zoomMin: PropTypes.number,
   }),
   parent: PropTypes.object,
@@ -94,4 +87,4 @@ MapSelectAmsterdam.propTypes = {
   validatorsOrOpts: PropTypes.object,
 };
 
-export default MapSelectAmsterdam;
+export default MapSelectGeneric;
