+++ src/signals/incident/components/form/index.js
@@ -13,7 +13,7 @@ import TextInput from './TextInput';
 import TextareaInput from './TextareaInput';
 import MapInput from './MapInput';
 import MapSelect from './MapSelect';
-import MapSelectAmsterdam from './MapSelectAmsterdam';
+import MapSelectGeneric from './MapSelectGeneric';
 import CategorySelectRenderer from './CategorySelectRenderer/CategorySelectRenderer';
 
 export default {
@@ -26,7 +26,7 @@ export default {
   HiddenInput,
   MapInput,
   MapSelect,
-  MapSelectAmsterdam,
+  MapSelectGeneric,
   MultiTextInput,
   PlainText,
   RadioInputGroup,
