+++ src/signals/incident/definitions/wizard-step-2-vulaan/wegen-verkeer-straatmeubilair.js
@@ -166,7 +166,7 @@ export const controls = {
       legend_items: ['lichtmast', 'grachtmast', 'overspanning', 'gevel_armatuur', 'schijnwerper', 'overig_lichtpunt'],
       pathMerge: 'extra_properties',
     },
-    render: FormComponents.MapSelectAmsterdam,
+    render: FormComponents.MapSelect,
   },
   extra_straatverlichting_niet_op_kaart: {
     meta: {
@@ -303,7 +303,7 @@ export const controls = {
       zoomMin: 14,
       pathMerge: 'extra_properties',
     },
-    render: FormComponents.MapSelectAmsterdam,
+    render: FormComponents.MapSelect,
   },
   extra_klok_niet_op_kaart: {
     meta: {
